feat(diary): implement issue 241 Markdown reports and IME fixes
