--- .github/workflows/issue-241-apply.yml
@@ -1,45 +1,0 @@
-name: Apply issue 241 implementation
-
-on:
-  push:
-    branches:
-      - feat/issue-241-diary-ime
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    steps:
-      - name: Checkout branch
-        uses: actions/checkout@v4
-        with:
-          ref: feat/issue-241-diary-ime
-          fetch-depth: 0
-
-      - name: Apply and commit source changes
-        shell: bash
-        run: |
-          set +e
-          python3 scripts/apply_issue_241.py > issue-241-apply.log 2>&1
-          status=$?
-          set -e
-
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-
-          if [ "$status" -ne 0 ]; then
-            git add issue-241-apply.log
-            git commit -m "chore: capture issue 241 patch failure"
-            git push origin HEAD:feat/issue-241-diary-ime
-            cat issue-241-apply.log
-            exit "$status"
-          fi
-
-          rm -f issue-241-apply.log .github/workflows/issue-241-apply.yml scripts/apply_issue_241.py
-          git add -A
-          git commit -m "feat(diary): implement issue 241 markdown reports and IME fixes"
-          git push origin HEAD:feat/issue-241-diary-ime
